Move matcher inside service folder.
