+++ packages/unmock-core/src/index.ts
@@ -4,7 +4,7 @@ import { doUsefulStuffWithRequestAndResponse } from "./util";
 // top-level exports
 export { UnmockOptions } from "./options";
 export * from "./interfaces";
-export * from "./matcher";
+export * from "./service/matcher";
 export * from "./generator";
 
 export const util = {
